Cleanup on the license file
Some compliance scanning tools are considering Spring's license to be "Apache 2.0 Non-Standard" because it was copied "as-is" from Apache's page. The resulting text contains both the full license plus the alternative snippet meant to be used instead of the of full license, causing the scans to consider the text as an "Apache 2.0 Non-Standard".